Move the external_node script to puppet confdir
The external nodes script was originally located in the puppet-dashboard
config directory. This doesn't actually make sense. The external_nodes
script is purely a Puppet configuration thing, it just happens to point
at the console (as do many other Puppet config files). Having it live in
the puppet-dashboard directory raises complications about situations
where you're installing a master but puppet-dashboard (the app) is NOT
going to be installed. Why would you have a config dir for it then? It
makes much more sense to just put the external_nodes script in Puppet's
own confdir.